add one basic test
