+++ papermerge/core/dbapi.py
@@ -52,7 +52,8 @@ from .features.document_types.db.api import (
     get_document_types_without_pagination,
     get_document_types_grouped_by_owner_without_pagination,
     delete_document_type,
-    update_document_type
+    update_document_type,
+    document_type_cf_count
 )
 from .features.custom_fields.db.api import create_custom_field
 from .features.shared_nodes.db.api import (
@@ -86,6 +87,7 @@ __all__ = [
     "get_document_types_grouped_by_owner_without_pagination",
     "delete_document_type",
     "update_document_type",
+    "document_type_cf_count",
     "create_custom_field",
     "get_doc_versions_list",
     "get_doc_version_download_url",
