+++ papermerge/core/features/document_types/tests/test_document_type_dbapi.py
@@ -20,3 +20,13 @@ async def test_get_document_types_grouped_by_owner_without_pagination(
     )
 
     assert len(results) == 2
+
+
+async def test_document_type_cf_count(db_session: AsyncSession, make_document_zdf, user):
+    zdf_doc_instance: orm.Document = await make_document_zdf(title="ZDF Title", user=user)
+
+    cf_count = await dbapi.document_type_cf_count(
+        db_session, document_type_id=zdf_doc_instance.document_type_id
+    )
+
+    assert cf_count == 3
